Prepare executables for packaging
Move existing executables to suit packaging structure while
retaining present functionality.
